Debug/fix airship/charts github repo mirroring
OpenDev->GitHub repo mirroring in not working for the charts repo,
and the logs for the failure are no longer available.

For this patchset I've re-encrypted the SSH key, in case there was
an issue with it previously; and otherwise, merging this patchset
(and re-running the mirror job) should produce some fresh logs
to debug.

Change-Id: I02fd44a051ff008f9562512cc8c710759daa1f5b