Update GitHub secret
This change updates the GitHub secret, used by our repository mirroring
job, with a re-encrypted version of the key. At the moment, the
mirroring job is not successful due to an invalid private key format.
The hope is that this patch will restore functionality; however, the key
cannot be tested until this change is merged.

Change-Id: Ib7dc16015d81206eb69174aa1258c7b6b1572ebb