Gerrit: ldap backed auth, verified votes, and webhook plugin
This PS updates Gerrit to use LDAP for credential management
permitting both easier CI by permitting HTTP basic auth for API
requests, and closer matching a real world config.

Verified labeled votes are also enabled and configured, for CI to
respond to.

Additionally the basis of using gerrit to trigger tekton pipelines
is validated.

As a rapid follow on to this, param passing from the webhook to the
trigger will need to be fleshed out.

Change-Id: Id135c71c7f94df14e5583b5ceb7db25688f89dfe
Signed-off-by: Pete Birley <pete@port.direct>