Improve the possibility to inherit zuul jobs
So far it takes some effort and workaround to inherit zuul jobs,
airship-jarvis, in particular. Gate scripts are running from the
directory different from the opendev.org/airship/charts when inheriting.

Thus, adding a single variable (gate_scripts_relative_path) that is
responsible for setting a directory for the gate scripts should solve
the issue and improve the inheritance.

Change-Id: Icc06211a8d5a0814f5e902ebee97d440d3d2943a