feat(entrypoint) Create entrypoints
This patchset moves all of the Tekton task script sections to use
a defined entrypoint instead of an ansible-playbook command. This is
a step to keeping all business logic out of the tekton tasks, i.e. the
tekton tasks should work with any standard container given, not just one
using Ansible.

Change-Id: I5e106a8a75b79c0c2948cda638fbe532fd12fae3