fix(standard-container/functional): create namespace for Helm release
Before Helm would fail to deploy the release if the namespace was
missing in the cluster. Now Helm will create the namespace if missing.

This isn't hit when the target cluster is the same cluster running the
tekton pipelines. This issue is only hit when the kubeconfig provided
via jarvis.yaml is for another cluster which might not have the
namespace created.

Change-Id: I3fdc6b353a2af2a667884c3250108b9f6c6b9a02