fix(300-deploy-loki): modify Helm test script to support proxy
Before, invoking the Loki Helm test was skipped if the environment had a
proxy configured.

Now, the configmap for the test is modified with the proxy vars. The
Loki stach Helm chart does not provide any configuration around this, so
we can modify the configmap.

As of right now, the certificates are not needed for `apk add` and only
proxy vars are required.

Change-Id: If58d99555ed299b99bd9bda441856aac326d8379