fix(loki-stack): enable seeing logs for dev pipeline
This commit does a number of things to enable this functionality.

Regex for variables using $namespace were removed as this assumed
pipeline names were prefixed with the namespace, which is only true for
`jarvis-system`. Instead allow all matches based on the variable query.

Before the variables queried the Loki source, but Loki does not support
filtering like this [1]. The loki service with /loki endpoint can be
additionally added as a Prometheus source and this API is compatible
with filtering as desired.

With this combined, logs for development pipeline are viewable and
dropdowns now are limited in scope to what actually exists. Before all
taskruns were displayed in the dropdown, but now only the taskruns for
the given pipelinerun are selectable.

1 - https://github.com/grafana/grafana/issues/25205

Change-Id: I1b7094947bcad71c40425c3bdab22fabdcd45884