Add test for pre-defined image tags
This commit adds a test to verify that pre-defined image tags are
preserved during development pipeline runs. The deployment-flow configmap
is checked which would contain the original image tag and the image.json
is also checked which would contain the image tag that is actually used.
The test checks if the configmap tag matches the image.json tag, which it
should. The test skips if there is no pre-defined tag.

Change-Id: Ic7661dfc23bc92a460b3b0d5328a6f85b42a6ed4